refactor: move pivot subtotal data structure retrofit
